backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
google-generativeai==0.8.0
httpx==0.27.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
sse-starlette==2.1.0
pydantic==2.9.0
pydantic-settings==2.5.0
lxml==5.3.0
greenlet>=3.0.0
apscheduler>=3.10.0
